included some tests
